refactored streamlit viz modules into MVC pattern
